More work on issue #73

+++ trunk/ssf/main/src/com/sitescape/ef/util/FileUploadItem.java
@@ -147,7 +147,9 @@ public class FileUploadItem {
 	}
 	
 	public void delete() throws IOException {
-		FileHelper.delete(tempFile);
+		if(tempFile != null) {
+			FileHelper.delete(tempFile);
+		}
 	}
 	
 	private void setup() throws IOException {
